cex-broker-server-modularization spec impl
