+++ openspec/changes/cex-broker-server-modularization/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-06-05
