+++ openspec/changes/cex-broker-server-modularization/README.md
@@ -1,0 +1,3 @@
+# cex-broker-server-modularization
+
+Establish canonical `src/helpers/` and `src/handlers/` layout, extract `server.ts` helpers, deduplicate shared primitives, and shrink `server.ts` to gRPC wiring and dispatch only.
