+++ package.json
@@ -43,6 +43,7 @@
 		"lint:fix": "bunx biome lint --write",
 		"check": "bunx biome check",
 		"check:fix": "bunx biome check --write",
+		"check:server-lines": "bash scripts/check-server-line-budget.sh",
 		"copy:dts": "cpx \"build/**/*.d.ts\" ./dist/",
 		"copy:proto": "cpx \"src/proto/*.proto\" ./dist/proto",
 		"prepare": "bunx husky"
