+++ src/handlers/subscribe/index.ts
@@ -1,0 +1,1 @@
+export { createSubscribeHandler, type SubscribeDeps } from "./handler";
